+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- `showProductName` prop to `ProductGiftName` component. It allows the user to choose whether to show the skuname or the productname of the gift.
+- `nameType` prop to `ProductGiftName` component. It allows the user to choose whether to show the skuname or the productname of the gift.
 
 ## [0.2.1] - 2020-05-13
 ### Fixed